Izrada izgleda mobilne aplikacije
This repository showcases the mobile application I developed for an Upwork client, with the initial design and prototyping crafted using Figma. The project represents a seamless blend of design and development, highlighting my expertise in mobile app creation and UI/UX design. Explore the repository for a closer look at the code and design elements that brought this project to life.